requirements.txt
discord.py
bs4.py
http3.py